pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.12.0",
    "groq>=0.37.1",
    "isort>=7.0.0",
    "langchain>=1.1.3",
    "langchain-classic>=1.0.0",
    "langchain-groq>=1.1.0",
    "langchain-tavily>=0.2.13",
    "python-dotenv>=1.2.1",
]
